Working on Personal Space and Magician's spell
Play is the tester class. It will be different everytime. So use the
play class from other verisons if need
03/31/13 - Sriram

Character
added Shape personalSpace variable - the shape covers the character’s
image
04/06/13

Character
int direction - returns the direction of the character. 12 is up. 6 is
down. Like a clock
Changed name of personalSpace to personal
updatePersonal() - changes the x and y coordinates, height, width of
the shape
drawPersonal() - draws the personal on the screen.
Everything is commented
Magician
Everything is commented.